// src/app/interfaces/Filtro.ts
export interface Filtro {
  nome?: string;
  cpf?: string;
  login?: string;
  idStatus?: number;
  periodoNascimentoInicial?: string;
  periodoNascimentoFinal?: string;
  periodoInsercaoInicial?: string;
  periodoInsercaoFinal?: string;
  periodoAlteracaoInicial?: string;
  periodoAlteracaoFinal?: string;
  idFaixaEtaria?: number;
}